First version of a subdomain checker and cleaner.
With this script you can check a list of subdomains with a list of out of scope subdomains and get
rid of every domain that is out of scope, after that, it will check in scope subdomains for a web
server running on it, and finally it will check for all web subdomains that are api links and create
a text file for each one of them.

NOTE: Want to add sublist3r as a first step to automate the process even more.